Convert sensor data from ints to floats
Originally all fields were ints, but now with the addition of
distance and speed there are also floats. Rather than use
separate float / int sensor values it was assumed simpler to
convert them all to floats and update the tests.

This also adds a test for the DistanceMeters field in the the
TCX file, but not for speed which will probably be removed
later.